start refactor hub page
